% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_trait_moments.R
\name{com.trait.moments}
\alias{com.trait.moments}
\title{Community Trait Moments}
\usage{
com.trait.moments(my.sample, traits)
}
\arguments{
\item{my.sample}{A community data matrix.}

\item{traits}{A trait file with species names as row names and trait as columns.

-\code{ com.mean } The mean trait value of the community
-\code{ com.sd } The standard deviation of the trait values in the community
-\code{ com.skew } The skewness of the trait values in the community
-\code{ com.kurt } The kurtosis of the trait values in the community}
}
\value{
a data.frame
}
\description{
Quantifies the moments of the community trait distribution
}
\examples{
data("lefse_sample")
data("lefse_traits")
com.trait.moments(lefse_sample, lefse_traits)
}
\references{
Swenson, N.G. 2014. Functional and Phylogenetic Ecology in R. Springer UseR! Series, Springer, New York, New York, U.S.A.
}
\seealso{
\code{\link{com.trait.weighted}}
}
\author{
Nathan G. Swenson
}
