% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fmntd_a.R
\name{Fmntd.a}
\alias{Fmntd.a}
\title{Functional NND with abundance}
\usage{
Fmntd.a(dist.mat, my.sample)
}
\arguments{
\item{dist.mat}{A trait distance matrix.}

\item{my.sample}{A community data matrix.}
}
\value{
The abundance weighted mean nearest functional neighbor distance for each community
}
\description{
Quantifies the functional mean nearest neighbor distance with abundance weighting
}
\examples{
data(lefse_sample)
data(lefse_traits)
Fmntd.a(as.matrix(dist(lefse_traits)), lefse_sample)
}
\references{
Swenson, N.G. 2014. Functional and Phylogenetic Ecology in R. Springer UseR! Series, Springer, New York, New York, U.S.A.
}
\seealso{
\code{\link{Fmntd}}
}
\author{
Nathan G. Swenson
}
