% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fmpd_a.R
\name{Fmpd.a}
\alias{Fmpd.a}
\title{Functional MPD with abundance}
\usage{
Fmpd.a(dist.mat, my.sample)
}
\arguments{
\item{dist.mat}{A trait distance matrix.}

\item{my.sample}{A community data matrix.}
}
\value{
The abundance weighted mean pairwise distance for each community
}
\description{
Quantifies the functional mean pairwise distance with abundance weighting
}
\examples{
data(lefse_sample)
data(lefse_traits)
Fmpd.a(as.matrix(dist(lefse_traits)), lefse_sample)
}
\references{
Swenson, N.G. 2014. Functional and Phylogenetic Ecology in R. Springer UseR! Series, Springer, New York, New York, U.S.A.
}
\seealso{
\code{\link{Fmpd}}
}
\author{
Nathan G. Swenson
}
