% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fmpd.R
\name{Fmpd}
\alias{Fmpd}
\title{Functional MPD no with abundance}
\usage{
Fmpd(dist.mat, my.sample)
}
\arguments{
\item{dist.mat}{A trait distance matrix.}

\item{my.sample}{A community data matrix.}
}
\value{
The presence-absence weighted mean pairwise distance for each community
}
\description{
Quantifies the functional mean pairwise distance with no abundance weighting
}
\examples{
data(lefse_sample)
data(lefse_traits)
Fmpd(as.matrix(dist(lefse_traits)), lefse_sample)
}
\references{
Swenson, N.G. 2014. Functional and Phylogenetic Ecology in R. Springer UseR! Series, Springer, New York, New York, U.S.A.
}
\seealso{
\code{\link{Fmpd.a}}
}
\author{
Nathan G. Swenson
}
