% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fsntd.R
\name{Fsntd}
\alias{Fsntd}
\title{Functional standard deviation of the NND}
\usage{
Fsntd(dist.mat, my.sample)
}
\arguments{
\item{dist.mat}{A trait distance matrix.}

\item{my.sample}{A community data matrix.}
}
\value{
The presence-absence weighted standard deviation of the nearest functional neighbor distance for each community
}
\description{
Quantifies the functional standard deviation of the nearest neighbor distance with no abundance weighting
}
\examples{
data(lefse_sample)
data(lefse_traits)
Fsntd(as.matrix(dist(lefse_traits)), lefse_sample)
}
\references{
Swenson, N.G. 2014. Functional and Phylogenetic Ecology in R. Springer UseR! Series, Springer, New York, New York, U.S.A.
}
\author{
Nathan G. Swenson
}
