% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefse-package.R
\docType{package}
\name{lefse-package}
\alias{lefse}
\alias{lefse-package}
\title{lefse: Phylogenetic and Functional Analyses for Ecology}
\description{
Utilizing phylogenetic and functional information for the analyses of ecological datasets. The analyses include methods for quantifying the phylogenetic and functional diversity of assemblages.
}
\author{
\strong{Maintainer}: Nathan G. Swenson \email{nswenson@nd.edu}

Other contributors:
\itemize{
  \item Paul Efren Santos Andrade \email{paulefrens@gmail.com} (\href{https://orcid.org/0000-0002-6635-0375}{ORCID}) [contributor]
}

}
\keyword{internal}
