% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefse_sample.R
\docType{data}
\name{lefse_sample}
\alias{lefse_sample}
\title{Community data matrix for lefse}
\format{
A data frame with 5 observations on the following 10 variables.
}
\source{
Swenson, N.G. 2014. Functional and Phylogenetic Ecology in R. Springer UseR! Series, Springer, New York, New York, U.S.A.
}
\usage{
data("lefse_sample")
}
\description{
Community data matrix for lefse package. Matrix contains 10 species and 5 communities
}
\keyword{datasets}
