% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefse_traits.R
\docType{data}
\name{lefse_traits}
\alias{lefse_traits}
\title{Example trait dataset for lefse}
\format{
A data frame with 10 observations on the following 3 variables.
}
\source{
Swenson, N.G. 2014. Functional and Phylogenetic Ecology in R. Springer UseR! Series, Springer, New York, New York, U.S.A.
}
\usage{
data("lefse_traits")
}
\description{
Example trait dataset for lefse package. The trait matrix contains 10 species and 3 traits
}
\keyword{datasets}
