% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_range.R
\name{trait.range}
\alias{trait.range}
\title{Range of trait values in a community}
\usage{
trait.range(my.sample, traits)
}
\arguments{
\item{my.sample}{A community data.matrix.}

\item{traits}{A trait matrix.}
}
\value{
The range of trait values for each community.
}
\description{
Quantifies the range of trait values present in a community
}
\examples{
data(lefse_sample)
data(lefse_traits)
trait.range(lefse_sample, lefse_traits)
}
