% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_trait_weighted.R
\name{com.trait.weighted}
\alias{com.trait.weighted}
\title{Quantifies the moments of the community trait distribution weighted by abundance}
\usage{
com.trait.weighted(my.sample, traits)
}
\arguments{
\item{my.sample}{A community data matrix.}

\item{traits}{A trait file with species names as row names and trait as columns.}
}
\value{
a data.frame
}
\description{
Quantifies the moments of the community trait distribution weighted by abundance
\itemize{
\item \code{com.weighted.mean} The abundance weighted mean trait value in the community
\item \code{com.weighted.sd} The abundance weighted standard deviation of trait values in the community
}
}
\examples{
data("lefse_sample")
data("lefse_traits")
com.trait.weighted(lefse_sample, lefse_traits)
# Sourse SDMTools package
wt.var <- function(x,wt) {
 s = which(is.finite(x + wt)); wt = wt[s]; x = x[s] #remove NA info
 xbar = weighted.mean(x,wt) #get the weighted mean
 return( sum(wt *(x-xbar)^2)*(sum(wt)/(sum(wt)^2-sum(wt^2))) ) #return the variance
}

wt.sd <- function(x,wt) {
 return( sqrt(wt.var(x,wt)) ) #return the standard deviation
}
}
\references{
Swenson, N.G. 2014. Functional and Phylogenetic Ecology in R. Springer UseR! Series, Springer, New York, New York, U.S.A.
}
\author{
Nathan G. Swenson
}
